servicio-reconocimiento-senas/requirements.txt
google-genai
fastapi
uvicorn[standard]
pillow
python-multipart